requirements.txt
google-generativeai
python-docx
python-dotenv
pyspellchecker
json5
fastapi
uvicorn
python-multipart
pydantic
openai
httpx
supabase
PyJWT
PySide6